refactor(TemplateLibrary): refactor TemplateLibrary into container connected to store
